// web/src/api/types.gen.ts
export type PatchApiStoresByIdData = {
    body: {
        name?: string;
        location?: string;
        manager?: string;
        status?: 'active' | 'inactive';
    };
    path: {
        id: number;
    };
    query?: never;
    url: '/api/stores/{id}';
};